Sketch out an AWS CloudFormation based compilation of the vote50 app
This change adds a .mu/ directory to the vote50 app.  Inside are the "expected"
contents that the Mu compiler should output.  Eventually we will validate against
this that the (currently non-existent) compiler generates the correct thing.

This isn't complete, although I can manually create bits and pieces of the stacks
using CloudFormation.  More files will come; for example, the code packages.

In a nutshell:

    .mu/                                # Shared files agnostic to the cloud target
        aws-native-cf/                  # Files specific to the AWS CloudFormation target
            app.template.yaml           # The CloudFormation stack for the overall app
            services/                   # A directory containing service-specific stacks
                voting.template.yaml    # The CloudFormation stack for the VotingService
        services/                       # Service files agnostic to the cloud target
            voting.proto                # The gRPC definition for VotingService's API
        mu.yaml                         # The Mu package manifest (minimal for now)