Don't capnext if the first rune is unprintable
If the first rune is unprintable, then we don't want to go ahead and force
capitalization on the next character.  (Unlike any other non-first rune,
where of course we do.)  In the case of a first rune, we want to let the
current default based on the pascal parameter take charge.