Don't rebind properties
There are two cases to consider in the case of a stack's bound properties.
First, it was a stack we didn't need to construct.  This is the case for
built-in primitives.  In that case, we must rebind each service uniquely.
Second, it was an imported stack, which by definition we had to construct.
In this case, we don't need to rebind the properties -- we already did so
-- and can just reuse them in the service's own bound properties.

I am pondering a better way of representing this and will probably do this
soon.  The concept of unconstructed vs. constructed types should be unified
between the built-in types and imported ones.  (Kind of like generics.)
But, I still need a little bit more time prototyping before I make up my
mind what a better way to represent all of this might look like.