Map Route's internet gateway as GatewayId
The route resource is a bit funny, in that it lets you jam in either
an internet gateway ID or a VPC ID as the GatewayId property, whereas
all of the other target kinds get their own property.  Our stack type
offers stronger typing than this (using service references), but we
need to map the resulting ref strings correctly.