Fix references to each
In an older version of this doc, `if` was called `when`, and `for`
was called `each`, to encourage a more declarative feel.  I renamed
them back to their familiar forms.  I missed a few spots.  But I also
thought at least documenting this heritage would be useful, hence
this descriptive checkin comment.