Add IsInst and TypeOf expressions
This adds two expressions we will need to support many type test patterns.

IsInst simply checks an expression for castability to a target type; it evaluates
to a boolean.

TypeOf fetches the runtime type of a given expression and evaluates to its name.
Right now, this is just a string token, rather than some richer RTTI object.