Export a top-level list of modules
The way ECMAScript and Python work is that each file maps to a module.  There
isn't really an "official" notion of a submodule; instead, directories on the
import path are conventionally used as submodule names.  Aggregate submodules
can be achieved by convention with re-exports in ECMAScript, however this is an
informal notion.  We could keep the current structured submodule notion and require
that all languages map to it, however, it is just as easy to have a flat list of
top-level modules indexed by name.  This change rearranges the AST accordingly.