Implement variable declaration lowering
This is an initial cut at implementing variable declaration lowering.

The lowering is slightly different for module properties and local variables,
since the former requires that code go into the module initializer, while the
latter can mostly happen "inline" as the local variable is bound.

There are some incomplete parts in here -- like lowering type names -- that
are dependent on a few upcoming changes.