Add a baseline for the empty test
This adds a baseline for the empty module test, and it passes.
Unfortunately, because module names are absolute file paths, this
baseline file has my machine's specific path embedded in it.  I want
to lock these in so they are passing as I make progress.  Sometime
later this morning I'll fix the module naming to be "relative" and
more Mu-friendly, and this issue will go away.