Add basic Mu metadata parsing support
This change adds discovery of Mu metadata in the form of a Mu.json file.
This allows us to eliminate the "TODO" in place of a real package name in
the output ASTs, so our baselines are adjusted accordingly.  This is part
of marapongo/mu#42 -- implementation of real module and package names.