Shorten module names
Per marapongo/mu#42, this change shortens module names to be more "Mu-like"
and to reflect the relative module paths.  There is still awkwardness if there
are any "up and over" references, i.e. using "..", but this handles the 95%
case.  We still don't mark entrypoints, but this gets us one step closer.