Map new expressions
This change adds a new AST node kind to represent "new" expressions.
Originally I had thought we could simply represent these as calls to
a constructor, however, there is a sizeable semantic difference and so
it's better for this node kind to be distinct; for example, we will
most likely permit allocation of objects without explicit constructors.

In addition, I've added a "complex" scenario test, which is just a
Point class.  It has a constructor, two properties, and an instance
method that performs arithmetic and allocates a new instance.  It works!