Add real diagnostics
This change adds a set of true diagnostics constructs, underneath the
new `diag` module.

This includes projecting Mu-specific errors as real diagnostics in a
way that is unified with TypeScript errors.  The only difference, of
course, is that Mu errors tend to happen in later passes.  But this is
not necessarily always the case.

As part of this, I've rearranged the compiler passes to present a
simpler interface to users of the compiler API (currently just the CLI
and test harness, but it's just a library, so anybody can use it).

Namely, there are three phases:

    1. Compilation is the overall process of taking an input path
       and driving the entire compilation process, yielding a set of
       diagnostics and, ideally, a final MuPackage at the end of it.

    2. Script compilation is the process of driving the front-end
       compiler -- in this case TypeScript -- yielding a set of Mu
       diagnostics and, if that went well, the script's AST.

    3. Transformation is the process of taking the script output and
       lowering it into the final MuPackage form.

Most people will deal with 1, blissfully unaware of the presence of
independent 2 and 3 phases.  We choose to keep them distinct for
white box testing and for future scenarios we have yet to envision.