Improve line-by-line error reporting in test failures
This change improves the line-by-line error reports when a test fails,
both in the diagnostics messages and AST tree baseline comparisons.
Previously, the test would fail at the first occurrence of a mismatch,
often with little information about the context.  In this change, the
test will now fail with a full dump of the comparison.

For instance, here is the before for a test failure in a new case I'm adding:

     AssertionError: Expected tree line #1 to match
           + expected - actual

                 -     "name": "iface_1",
                 +     "name": "class_1",

In comparison, here is what the test failure looks like now:

      AssertionError: Expected trees to match; 6 did not
      + expected - actual

      -1:     "name": "iface_1",
      -10:                 "I": {
      -14:                         "ident": "I",
      -19:                                 "column": 10
      -23:                                 "column": 11
      -29:                     "interface": true,
      +1:     "name": "class_1",
      +10:                 "C": {
      +14:                         "ident": "C",
      +19:                                 "column": 6
      +23:                                 "column": 7
      +29:                     "abstract": false,