Map MuJS interface types to MuIL
This change maps MuJS interface types to MuIL, including property and
method declarations.  The mapping is fairly straightforward, and similar
to the existing class lowering, except for some minor AST differences in
the TypeScript representations.  We emit interface types as classes with
the interface bit set, and all method declarations being abstract.