Discard import declarations
At the moment, we do not emit bound node information yet.  (See
marapongo/mu#46).  When that day comes we will need to decide whether
import declarations are needed or, as I suspect, they can continue to
be discarded safely because all dependency information is available
on the bound nodes themselves.  I'm leaving in a TODO in there just as
a reminder to resolve this at that time.